fix cho product detail and cart, order vs checkout van loi

+++ src/main/java/com/example/fruitmarket/service/CartService.java
@@ -7,9 +7,9 @@ import org.springframework.stereotype.Service;
 public interface CartService {
     Cart getCart();
 
-    void addToCart(Long productId, Long variantId, int qty);
+    void addToCart(Long productId, Long variantId, double qty);
 
-    void updateQuantity(Long productId, Long variantId, int qty);
+    void updateQuantity(Long productId, Long variantId, double qty);
 
     void remove(Long productId, Long variantId);
